test(packaging): split runtime resource checks
